Add version discovery information to the /v1 endpoint
Currently we do not provide enough information for keystoneauth
to work on. This makes version discovery impossible e.g. in a standalone
case when the versioned endpoint is supplied by a user.

This change adds the required versioning information based on the actual
keystoneauth implementation and the in-progress API SIG specification:
https://review.openstack.org/#/c/459710/

Note that this change is not microversioned, as it is used to discover
available microversions for the versioned endpoint.

Change-Id: Ibb1e8a1c8973a4742b7ba508dc84de577cb967c9
Story: #2003567
Task: #24859