Fix doc builds for ironic
It seems that the wsme sphinx extension was concealing
issues in our doc builds, where formatting was not exactly
correct nor parsed as we would always expect as we have had
a tendency to fully document data structures as if we are
typing code.

Upon disabling the extension, the prior build failures due to
duplicate labels disappeared and other issues preventing the
docs from building successfully were raised.

It also seems that we never actually used the extension as
documented https://wsme.readthedocs.io/en/latest/document.html

$ grep -ir "wsme:root" ./ironic
$ grep -ir "wsme:type" ./ironic
$ grep -ir "wsme:attribute" ./ironic
$ grep -ir "wsme:service" ./ironic
$ grep -ir "autotype" ./ironic
./ironic/objects/fields.py:class FlexibleDictField(object_fields.AutoTypedField):
./ironic/objects/fields.py:class MACAddressField(object_fields.AutoTypedField):
$ grep -ir "autoattribute" ./ironic
$ grep -ir "autoservice" ./ironic
$ grep -ir "autofunction" ./ironic
$

Co-Authored-By: Dmitry Tantsur <dtantsur@redhat.com>
Change-Id: Ib5bbd9274a4098022fba69f0e4998485d2ee2396
Story: 2003020
Task:  23043