Honors return value from BIOS interface cleansteps
When a BIOS interface clean step is executed the return value is ignored.
If it is asynchronous, the return value which is CLEANWAIT state is ignored
and the cleanstep is marked as complete immediately. The current change
honors the return value and moves the node to CLEANWAIT state accordingly.

Change-Id: I3a3915116286326316ee9a55333c046f729c1a08
Story: #2003750
Task: #26438