Completely remove support for deprecated Glance V1
Also removes support for having several API versions for the Image
API. We do not do it for Neutron and Swift. And if we ever support
the (hypothetical) Glance V3 API, we better do it transparently,
not with a configuration option (maybe using openstacksdk).
Cleans up Glance V1 specific features like image.is_public or
a separate image.properties field.

The code should be consolidated into fewer modules in a follow-up.

Story: #1670423
Task: #26830
Change-Id: I08a01eadbe8a2dec4c02674bed0391846632ef06