Avoid race with nova on power sync and rescue
Nova power sync can see an instance off during the rescue test,
and as a result decide that the power must remain off. As such,
we kind of need to fake it for a moment until we actually turn
power back on.

This is done in ironic to help make a minimal behavior fix as
ironic will return the power state to ``POWER_ON``.

Change-Id: Ib8210062560d6f5c7061caf1fe19f11290578e51
Story: 2003990
Task: 26934