Enhanced checksum support
This patch adds the ability to pass new hash fields os_hash_algo
and os_hash_value to IPA.

Glance will not re-compute hash values for existing images (they only
update these fields for newly uploaded images), so the os_hash_algo
and os_hash_value can be None.

In the case of direct interface with http provisioning, there
are two hash computing, one is for md5 checksum, and the other is
for new os_hash_algo (if it exists). If the os_hash_algo is
configured to md5, we'll bypass and wouldn't pass os_hash_algo to
IPA, because it's a waste of resource.

Change-Id: Iff72194787561b936efe3572e2b7a70217165a82
Story: 2003938
Task: 26845