Remove the xclarity deprecation
The lenovo team has setup third party CI which is reporting
into gerrit the job status. While their CI is still early in
the tests it executes, the log they provided to demonstrates
that they were able add a node to be managed via xclarity
and obtain the current boot devices.

They are diligently working on making logging available,
and as I understand it they will soon be running a deployment
scenario.

As such, I'm confident that we can go ahead and remove the
deprecation notice and flags.

Change-Id: I556c7ab42a74ebaa1f65642f2f18d23885de5ce3
Story: 2003097
Task: 27648