Allow disabling instance image cache
We document that this can be disabled by setting the
instance_master_path config  to "<None>",
but don't actually support it in code. oslo.config doesn't actually
translate that value to the Python None as we expect.

Allow disabling the cache by setting the config to the empty string, as
in "instance_master_path=". This doesn't make sense as a directory to
use as a cache anyway, so it shouldn't break anyone.

Conflicts:
	ironic/drivers/modules/deploy_utils.py
Due to class InstanceImageCache being in iscsi_deploy.py instead of
deploy_utils.py.

Change-Id: I1bb62d55e3e18272fd5da355d63fd2c11a033acd
Story: 2004279
Task: 27829
(cherry picked from commit 0498e2982901a586f19c9c8412c596bad0cf8004)