Expose conductors: db and rpc
This patch lays some ground work around db and rpc to provide
conductors information from the API.

Changes in the db api and Conductor object is used to support
the implementation of /v1/conductors. Adds an argument
"online" to Conductor.get_by_hostname, so that we can get
the conductor object from database even it's not online,
this is required for the implementation of /v1/conductors/{hostname}.

Adds a new interface get_conductor_for() to get the hostname
of the conductor which is servicing the given node, it will
be used for the implementation of /v1/nodes* endpoints, as well
as listing nodes by given conductor.

Story: 1724474
Task: 28064

Change-Id: I39a7a47c5ae649f6c3200e772a9357023f21a7c4