Slim down grenade jobs
As we can't extend our grenade job timeout past 3 hours,
and we have some fairly log running jobs, that essentially
means we need to find time that is not absolutely necessary
in the entire test run in order to reduce load.

Disabling cleaning for the grenade tests as well as
setting a regex for the basic server ops as opposed to
basic network ops which can take as much as 3x the amount
of time that the basic test takes.

Change-Id: Ib795e6dc3ad1f484f79bee81a5bcfdb8c3434fa1