Extend job build timeout
On slower nodes, sometimes Ironic can take up to ten
minutes for the deployment to complete. The downside is
that the tempest build timout is set to ten minutes.

In order to reduce the number of erronious failures
and rechecks, extending the timer to 12 minutes which
should provide enough time for the deployment to complete
to tempest's satistfaction.

Change-Id: Ie883d641d0951cb148e8cfb812b9607fa77ddb4b