Temporarily mark multinode job non-voting
The multinode job is failing again in stable branches pike
and queens. In queens the job is already non-voting, marking
it non-voting in pike as well until we are able to discover
the root cause and have a fix.
This will allow contributors to keep working on the pike branch
until we resolve the job issues.

Change-Id: I2f72b042996e3667315f5b79c132e1cb404b2639