Fix IPv6 iPXE support
Adds transmission of an `ipxe6` tag to neutron which is required
for neutron to appropriately match and supply configuration to
hosts over DHCPv6 when iPXE is in use.

This is because neutron prepends options based upon its
perception of the IP version being used and the detection of
iPXE actively being used varries between IPv4 and IPv6.

Story: 2004502
Task: 28221

Change-Id: If68926fd215f707394e5d6a0af9daa9232216abe