Allocation API: allow skipping retries in TaskManager
To reduce the number of times the API returns HTTP CONFLICT we retry
acquiring an exclusive lock in TaskManager. However, for the allocation
API it's not desired since we usually have several nodes to work with,
and another node can be picked without waiting for the current one
to become available. The new "retry" argument to TaskManager allows
disabling the retries.

Change-Id: Ia8267b1baa8b671a602e7fc402b7ef083e2ac2ed
Story: #2004341
Task: #28722