Add a high level vision reflection document
As part of the TC effort to raise visibility, we want to add
a short high level document which attempts to detail the areas
that could use additional focus where the project as a whole
does not seemingly align with the rest of the OpenStack Cloud
vision.

Some aspects really don't apply, but these seem to be valid areas
and it is important that we write these sorts of things down.

Change-Id: I280c73b8cabcd214d791293d25aef57216464ec7