Fix test for 'force_persistent_boot_device' (i)PXE driver_info option
The string used for the 'force_persistent_boot_device' field in the
node's driver_info is split by an additional blank and hence doesn't
set the option to be tested. In addition, the test condition tests a
mostly independent call (which also explains why the test succeeds
despite the incorrect option).

This patch removes the incorrect test and introduces new ones to
test the signature of prepare_ramdisk() which is the function in
which 'force_persistent_boot_device' is actually used.

Change-Id: I0b6a298837306fb8c6a46c13612b003dc62a7850