Add systemd unit for vbmcd in devstack
Besides managing vbmcd from systemd, this patch fixes a bug
in vbmcd handling.

The bug was that devstack runs ironic plugin under user `stack`, but
preserving superuser environment by means of `su -p`. That makes
`vbmcd` expanding tilda into `/root`, attempting to write into that
directory and failing miserably.

On top of that, all this happens when `vbmc` tries to invoke `vbmcd`
under the hood to preserve backward compatibility of the `vbmc`
behaviour before `vbmcd` was introduced.

This patch:

* adds systemd unit file for vbmcd
* invokes `vbmcd` explicitly before everything else and
* does that before becoming root and `stack` again

The `vbmc` tool can still run in a broken environment
for as long as it does not depend on the config file.

Change-Id: I6670ea50d126065d0472e558943890d4f9313083