Fix pysendfile requirement marker
Currently, any patch that changes one of the requirements files is
failing the requirements-check job, with the following error:

Requirement(package='pysendfile', location='', specifiers='>=2.0.0',
markers='', comment='# MIT', extras=frozenset()) 'markers': '' does not
match "sys_platform!='win32'"

Could not find a global requirements entry to match package {}. If the
package is already included in the global list, the name or platform
markers there may not match the local settings.

The problem is that a "sys_platform!='win32'" marker was added [1] to
the pysendfile requirement in the requirements repo, which needs to be
propagated to the requirements of all repos using the package.

This change adds the required marker to requirements.txt.

[1] https://review.openstack.org/#/c/639084/

Change-Id: Ibd7656cbcfbf204930f9c8562f79ab92bd4aefb5
Story: 2005096
Task: 29697