Software RAID: Trigger grub installation on the holder disks
Ironic triggers the bootloader installation only for partition
images (as whole disk images should come with the bootloader).
When a software RAID device is used for deployment, however, the
bootloader needs to be installed on the holder disks by the IPA,
so we need to detect and trigger this.

Change-Id: I6732bfbcabfa1c74aa5ab6fa990e977589a50b92
Story: #2004581
Task: #29105
Depends-on: https://review.opendev.org/639390