Temporary marking ironic-standalone non-voting
The ironic-standalone job is failing because the most recent IPA
image has only python3 and ansible uses python 2 as default
interpreter, resulting in a /usr/bin/python not found error.
Marking ironic-standalone non voting and removing it from the
gate until we can run ansible with python3 on remote machines.

Change-Id: I56768d7c8a577f267172de02d77e2cc2ead785ad