Fix idrac Job.state renamed to Job.status
In python-dracclient 0.1.0, the Job.state field was renamed to
Job.status. The old name was left for backwards compatibility, but was
removed in python-dracclient 2.0.0.

This change uses the new field name, to avoid an AttributeError in
python-dracclient 2.0.0 onwards.

Also modifies the drac unit tests to use the Job, RAIDController,
PhysicalDisk and VirtualDisk objects from the dracclient library when
available.

Change-Id: I074347748e1ce0f749f47ce30d4cb123a70dc2b8
Story: 2005225
Task: 30007