Make metrics usable
Metric notifications were hard coded with a value indicating
"ipmi" for all hardware types and management interfaces which
offer sensor data to be transmitted as a metric notification.

Also adds a node_name field to allow processors to leverage
the node name as opposed to needing ironic API or database access
to construct enough data to make metrics notifications usable.

Change-Id: I1116bd5812f3715bd7d6d1bf34df932f81702147