Fix idrac driver unit test backwards compat issue
Since commit 8bd3c531a0ef7282cc8a0daac9f71bb18b6a9532, the idrac driver
unit tests use real named tuple objects from the dracclient library, if
available. This added a coupling to a specific set of versions of that
library, since the correct set of fields for each object must be
specified.

This change uses the namedtuple._fields field to build a list of fields
appropriate to the library version. Missing fields are filled in with
None, and unsupported fields are ignored.

Change-Id: I5b9b64e5aa4fe96e056c8b010bf8a448c72f996f