Fix exception generation errors
This patch addresses some frequently failing tests with a few minor
changes related around how the exceptions are tested and how the
data is passed to the exceptions that are failing.

I've been able to duplicate the failing tests in CI while streaming
video on my desktop of all things. With these changes, I've been
unable to reproduce the failures.

This also fixes entries for DeployTemplate generation which were
noticed as well, however the failures were not observed in CI.

And a missing i18n tag was added into the pxe code, where it was
previously missing.

Change-Id: I719fa12340b51c55c0441df572ee7e3b33b1910c