ipmi: Ignore sensor debug data
When the conductor has debugging enabled, that command is
passed to ipmitool to enable debugging of other commands.

However, this means tons of extra data is dumped as part of
the sensor data collection for ironic, which breaks string
parsing and ultimately metrics collection.

Since we can identify these lines, lets ignore them.

Change-Id: Ife77707210f8289d8f2e0223fb9ee1909d798546
Story: 2005332
Task: 30267