Revert "Fix OOB introspection to use pxe_enabled flag in idrac driver"
This fix is inconsistent with the stable branch policy [1]. It requires
python-dracclient version 1.5.0 and greater; however,
driver-requirements.txt specifies version 1.3.0 and greater can be used
on this branch.

A new fix which does not require a version bump is being developed.

This reverts commit 828ec3ee397eb96427927ecd5f57d6f1c23a4af2.

The release note is not deleted from it, because the change has already
been released. A new one is added instead.

[1] https://docs.openstack.org/project-team-guide/stable-branches.html

Change-Id: I5afcbfb136a25d2e2101e3b09730f71b9b5d7c7d
Story: 2004340
Task: 27927