Revert "Fix OOB introspection to use pxe_enabled flag in idrac driver"
This fix is inconsistent with the stable branch policy [1]. It requires
python-dracclient version 1.5.0 and greater; however,
driver-requirements.txt specifies version 1.3.0 and greater can be used
on this branch.

A new fix which does not require a version bump is being developed.

This reverts commit feb3c5faea344eb149f2bc06accf012480336eec.

The release note is not deleted from it, because the change has already
been released. A new one is added instead.

[1] https://docs.openstack.org/project-team-guide/stable-branches.html

Change-Id: I5b6b37978cb820dfae2ad434f6a0d204fa1dd912
Story: 2004340
Task: 27927