Split ibmc power/reboot classes
We've seen some fun things in the past with
the test runner and things like sleep getting
mocked in the background resulting in some odd
behaviors when multiple tests are iterating
and performing calls.

Since that is the case, and we've seen the ibmc
reboot test case sporatically fail, lets try
splitting the reboot testing into its own test
class which results in a different runner thread
and some minor restructuring to limit variables.

Change-Id: I2051995cc7080ba42aedb57816ab8f63154c7f51