Place upper bound on python-dracclient version
The idrac hardware type requires python-dracclient versions greater than
or equal to 1.5.0 and less than 3.0.0.

This defines the upper bound to be less than 3.0.0 to make it easier to
backport to stable/stein. It was originally proposed against
stable/stein -- https://review.openstack.org/#/c/645956/ . Reviewers
asked that it be submitted against master first --
https://review.openstack.org/#/c/645956/1/driver-requirements.txt@12 --
and then backported to stable/stein.

The ironic master and stable/stein branches are compatible with
python-dracclient releases less than 3.0.0. Note that
python-dracclient's stable/stein branch was created off of tag 2.0.0.
The ironic stable/stein branch follows tag 12.1.0, which was created on
March 21. The python-dracclient tag 3.0.0 was created after that, on
April 4. The upper bound on master will be updated to less than 4.0.0 by
a follow-on change.

A related change has been proposed for stable/rocky --
https://review.openstack.org/#/c/651657/ . It will be backported to
stable/queens and prior after it merges.

Change-Id: I5f6d6be53aec9d72e8ba5cfa436f12e01966f871