Use the PUBLIC_BRIDGE for vxlan
The PUBLIC_BRIDGE parameter, for most multinode zuul jobs
defaults to br-infra in cases where native zuul configuration
is used.

However, this is not true for legacy jobs still using devstack-gate.

When jobs are invoked with devstack-gate, the bridge must be set
to match the devstack-gate scripting as zuul *does not* setup the
vxlan bridge for us. Devstack-gate does so.

Change-Id: I37bff3c4d930ada483690d0dea17cb9a64d51fd2