Place upper bound on python-dracclient version
The idrac hardware type requires python-dracclient versions greater than
or equal to 1.3.0 and less than 2.0.0.

This is not a backport of a similar change which merged on the master
and stable/stein branches [1]. That change specifies python-dracclient
versions greater than or equal to 1.5.0 and less than 3.0.0 are
required.

Subsequently, a change which merged only on master modified the bounds
to greater than or equal to 3.0.0 and less than 4.0.0 [2].

[1] https://review.opendev.org/#/q/I5f6d6be53aec9d72e8ba5cfa436f12e01966f871
[2] https://review.opendev.org/#/c/634903/

Change-Id: If5b0c5b69f0f8759136051072c1ff53c6d86217d