Option to send all portgroup data
By default our data model only sends port data for only ports
which we are attaching. That being said, depending on the switch
configuraiton, we need to transmit all port data for an ML2
driver to piece the universe together when port groups are
in use.

This patch adds a configuration option to allow deployments
to change the overall behavior so they can successfully
PXE boot machines with port groups.

Change-Id: I8d0f4e28efb05af704839f3b92cb3dfd48d57940
Story: 2005494
Task: 30595