Direct bridge to be setup
The multinode jobs are failing because the br-infra bridge
is not being setup. It is not being setup because somewhere
in the infra, the multi-node-bridge role is never established
as a requirement. Realistically, only ironic does this, but
we need to ensure it is present.

Change-Id: Ia1dcc68d7e75a931e01829ce23da36c978a70765
(cherry picked from commit bda02c43fb022a155157402764cd5a74478b8cf1)