Fix pyghmi path
Pyghmi was moved from openstack namespace to the x
namespace. As such the devstack plugin needs to match
up otherwise devstack fails in CI to clone as cloning
is blocked if not in the projects list... Which it is,
as x/pyghmi.

Change-Id: I87bdf65dd2dac189642d0af6def2d0ed155f3c00
(cherry picked from commit bb8101ca950819a4d55af18d57b53c15ef51c836)