devstack: configure rabbit outside of API configuration
If enabled, rabbit is required for all ironic services, so it makes
no sense to only configure it when API is enabled. This seems to
break multinode grenade, since `is_service_enabled ir-api` returns
false on the subnode. Try to configure the RPC backend in all cases
when IRONIC_RPC_TRANSPORT is "oslo" (not JSON RPC).

Also move the automated_clean option to the conductor where it belongs.

Change-Id: I6e046c6b6e1e4c0ca7bfd5fb9ce5892c0298fb78