Add an option to abort cleaning and deployment if node is in maintenance
We don't prevent cleaning to happen for nodes in maintenance mode.
However, cleaning cannot succeed in this case, as we disable processing
heartbeats. This change adds a new configuration option that will
cause such node to enter CLEAN FAIL on the first heartbeat.

The same is done for deployment and automated cleaning during providing.

Finally, elevate the log level for such heartbeats from debug to warning,
as it may be a sign of a problem (especially if the new option is off).

Change-Id: I9f3ee44f39c448eb2609c5989acd36e7da844ef4
Story: #1563644
Task: #9171