Add iDRAC RAID deploy steps
This change adds two new deploy steps to the idrac RAID interface -
apply_configuration and delete_configuration. These use the existing
RAID support in the idrac driver used for clean steps.

In order to make this work, the lifecycle controller job validation has
been modified to allow specification of a name prefix for jobs to check.
This is because configuring the node for PXE boot can result in creation
of a BIOS config job, which previously caused the validation to fail.
The RAID interface now only checks for existing jobs on the same RAID
controller, and so ignores the BIOS config job.

The disk space calculation has been modified to allow for virtual disks
that are pending deletion, as this is necessary to make the numbers work
when deleting existing virtual disks and creating new ones from the same
set of physical disks.

We also use the new deployment_polling flag in driver_internal_info to
signal that the RAID interface polls for completion of the deploy step.

Co-Authored-By: Shivanand Tendulker <stendulker@gmail.com>
Change-Id: I5803131fbdebce6f7896655a61a8fbdd4c1cd4a1
Story: 2003817
Task: 30004