Lower standalone concurrency to 3 from 4
The standalone job seems to be the most problematic CI job
in ironic's CI jobs at this point in time. While we've moved
to unsafe disk caching for the test VMs, there are times when
the devstack instance is still being crushed with IO wait for
multiple seconds.

This could be from processes in the VM or just a noisy neighbor
but first lets try to lower the concurrency to hopefully make
the job more reliable.

Change-Id: I4dfeb3407cfe4064a5d93a4ed3657677fa5237ef