Follow up to Option to send all portgroup data
Clarifies the code in add_ports_to_network() so that it is clear
that the ironic ports that are created in neutron are not necessarily
all pxe-enabled. Adds more comments to describe the behaviour if
non-pxe-enabled ports are created in neutron.

Fixes a typo in the corresponding release note.

This is a follow up to I8d0f4e28efb05af704839f3b92cb3dfd48d57940.

Change-Id: I94ff18bc137401ff4cd316755408c2099bb563cb
Story: 2005494