Fix get_boot_option logic for software raid
get_boot_option returns what the effective boot
option is for the scenario at hand for deployment.

That being said, it was never updated to account for the
presence of software raid. It will now return "local" as
the boot option when software raid is configured, as
Software RAID is only supported in local boot mode.

Change-Id: I2441cd60bb385648570e5994194e6fc6ff22025a